+++ lib/language_pickers.dart
@@ -2,6 +2,7 @@ library language_pickers;
 
 export 'languages.dart';
 export 'language_picker_dropdown.dart';
+export 'language_picker_dropdown_controller.dart';
 export 'language_picker_dialog.dart';
 export 'language_picker_cupertino.dart';
 export 'utils/utils.dart';